+++ sql/example_test.go
@@ -94,7 +94,7 @@ func Example() {
 	}
 
 	ctx := context.Background()
-	trManager := manager.New(trmsql.NewFactory(db))
+	trManager := manager.New(trmsql.NewDefaultFactory(db))
 
 	err = trManager.Do(ctx, func(ctx context.Context) error {
 		if err := r.Save(ctx, u); err != nil {