+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.0] - 2022-10-24
+
+### Added
+
+- timeout and canceling of transaction
+
+## [0.0.0] - 2022-10-20
+
+### Added
+
+- chained transaction
+
 ## [0.0.0] - 2022-10-17
 
 ### Added